+++ apps/noticiasweb3/src/App.jsx
@@ -34,6 +34,7 @@ import BienvenidoPage from './pages/HomePage.jsx';
 import NoticiasPage from './pages/NoticiasPage.jsx';
 import NoticiaDetailPage from './pages/NoticiaDetailPage.jsx';
 import BlogPostDetailPage from './pages/BlogPostDetailPage.jsx';
+import CanalPage from './pages/CanalPage.jsx';
 import ExtensionesPage from './pages/ExtensionesPage.jsx';
 import ExtensionDetailPage from './pages/ExtensionDetailPage.jsx';
 import PulserasRojasPage from './pages/PulserasRojasPage.jsx';
@@ -152,6 +153,7 @@ export default function App() {
         <Route path="/noticias" element={<Layout><NoticiasPage /></Layout>} />
         <Route path="/noticias/:slug" element={<Layout><NoticiaDetailPage /></Layout>} />
         <Route path="/blog/:slug" element={<Layout><BlogPostDetailPage /></Layout>} />
+        <Route path="/canal" element={<Layout><CanalPage /></Layout>} />
         <Route path="/iniciar-sesion" element={<Layout><IniciarSesionPage /></Layout>} />
         <Route path="/extensiones" element={<Layout><ExtensionesPage /></Layout>} />
         <Route path="/extensiones/:slug" element={<Layout><ExtensionDetailPage /></Layout>} />